refactor(editor): move editor to codex-ui
