+++ codex-ui/dev/Playground.vue
@@ -120,12 +120,14 @@
 
     <Heading :level="3"> Type Scale </Heading>
     <TypeScale />
+    <Heading :level="3">Editor.js</Heading>
+    <Editor :tools="{}" />
   </div>
 </template>
 
 <script setup lang="ts">
 import { ref } from 'vue';
-import { Button, Heading, Input, Field as FormField, Row, ButtonSize, ButtonStyle, Avatar } from '../src/vue';
+import { Button, Heading, Editor, Input, Field as FormField, Row, ButtonSize, ButtonStyle, Avatar } from '../src/vue';
 import TypeScale from './TypeScale.vue';
 
 const formFieldValue = ref('Heading');
