+++ codex-ui/package.json
@@ -41,6 +41,20 @@
     "vue-tsc": "latest"
   },
   "dependencies": {
+    "@editorjs/checklist": "^1.5.0",
+    "@editorjs/code": "^2.8.0",
+    "@editorjs/delimiter": "^1.3.0",
+    "@editorjs/editorjs": "^2.28.0",
+    "@editorjs/embed": "^2.5.3",
+    "@editorjs/header": "^2.7.0",
+    "@editorjs/image": "^2.8.1",
+    "@editorjs/inline-code": "^1.4.0",
+    "@editorjs/link": "^2.5.0",
+    "@editorjs/list": "^1.8.0",
+    "@editorjs/marker": "^1.3.0",
+    "@editorjs/raw": "^2.4.0",
+    "@editorjs/table": "^2.2.2",
+    "@editorjs/warning": "^1.3.0",
     "vue": "^3.4.16"
   }
 }
