+++ codex-ui/src/vue/index.ts
@@ -4,3 +4,4 @@ export * from './heading';
 export * from './input';
 export * from './row';
 export * from './avatar';
+export * from './editor';
